refactor: Restructure package module into separate files for improved maintainability
Split install, uninstall, refresh, extract commands, and tree view provider into src/packages/ directory.
Update extension.ts to register new modules, ensuring full functionality without omissions.
